fix: POS Update Order multiple requests.
