+++ components/ADempiere/Form/VPOS/Options/generalOptions.vue
@@ -1,7 +1,7 @@
 <!--
  ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
- Copyright (C) 2017-Present E.R.P. Consultores y Asociados, C.A.
- Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com www.erpya.com
+ Copyright (C) 2018-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
  This program is free software: you can redistribute it and/or modify
  it under the terms of the GNU General Public License as published by
  the Free Software Foundation, either version 3 of the License, or
@@ -9,11 +9,11 @@
 
  This program is distributed in the hope that it will be useful,
  but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  GNU General Public License for more details.
 
  You should have received a copy of the GNU General Public License
- along with this program.  If not, see <https:www.gnu.org/licenses/>.
+ along with this program. If not, see <https:www.gnu.org/licenses/>.
 -->
 
 <template>
@@ -206,14 +206,11 @@ export default defineComponent({
 
     const isShowProductsPriceList = computed({
       get() {
-        return root.$store.state['pointOfSales/point/index'].productPrice.isShowPopoverMenu
+        return root.$store.getters.getShowProductPriceList
       },
       set(isShowed) {
         if (!root.isEmptyValue(root.$route.query.pos)) {
-          root.$store.commit('showListProductPrice', {
-            attribute: 'isShowPopoverMenu',
-            isShowed
-          })
+          root.$store.commit('setIsShowListProductPrice', isShowed)
         }
       }
     })
